Optimize dataset hooks and LeftColumn visibility logic
